+++ src/main/java/harmonised/pmmo/gui/PrefsScreen.java
@@ -165,7 +165,7 @@ public class PrefsScreen extends Screen
     }
 
     @Override
-    public void render(int mouseX, int mouseY, float partialTicks)
+    public void render(int mouseX, int mouseY, double partialTicks)
     {
         renderBackground( 1 );
