+++ src/main/java/harmonised/pmmo/gui/Skill.java
@@ -2,8 +2,8 @@ package harmonised.pmmo.gui;
 
 public class Skill
 {
-	public float xp, pos, goalXp, goalPos;
-	Skill( float xp, float pos, float goalXp, float goalPos )
+	public double xp, pos, goalXp, goalPos;
+	Skill( double xp, double pos, double goalXp, double goalPos )
 	{
 		this.xp = xp;
 		this.pos = pos;
